move test files to better match the src directory
